src/utils ts 변경 완료
