--- src/utils/token.js
@@ -1,24 +1,0 @@
-import jwt, { decode } from 'jsonwebtoken';
-import { JWT_ACCESS_TOKEN_SECRET, JWT_REFRESH_TOKEN_SECRET } from './constants.js';
-
-export const generateTokens = (userId) => {
-  const accessToken = jwt.sign({ id: userId }, JWT_ACCESS_TOKEN_SECRET, {
-    expiresIn: '1h',
-    issuer: 'sprint-Mission4',
-  });
-  const refreshToken = jwt.sign({ id: userId }, JWT_REFRESH_TOKEN_SECRET, {
-    expiresIn: '6h',
-    issuer: 'sprint-Mission4',
-  });
-  return { accessToken, refreshToken };
-};
-
-export const verifyAccessToken = (token) => {
-  const decoded = jwt.verify(token, JWT_ACCESS_TOKEN_SECRET);
-  return { userId: decoded.id };
-};
-
-export const verifyRefreshToken = (token) => {
-  const decoded = jwt.verify(token, JWT_REFRESH_TOKEN_SECRET);
-  return { userId: decoded.id };
-};
